+++ sematic/resolvers/cloud_resolver.py
@@ -1,7 +1,8 @@
 # Standard Library
 import enum
 import logging
-from typing import Dict, List, Optional
+import pathlib
+from typing import Dict, List, Optional, Tuple
 
 # Third-party
 import cloudpickle
@@ -19,7 +20,11 @@ from sematic.db.models.factories import get_artifact_value
 from sematic.db.models.resolution import ResolutionKind
 from sematic.db.models.run import Run
 from sematic.resolvers.local_resolver import LocalResolver, make_edge_key
-from sematic.resolvers.resource_requirements import ResourceRequirements
+from sematic.resolvers.resource_requirements import (
+    KUBERNETES_SECRET_NAME,
+    KubernetesSecretMount,
+    ResourceRequirements,
+)
 from sematic.user_settings import SettingsVar, get_all_user_settings, get_user_settings
 
 logger = logging.getLogger(__name__)
@@ -249,11 +254,27 @@ def _schedule_job(
 
     node_selector = {}
     resource_requests = {}
-    if resource_requirements is not None:
+    volumes = []
+    volume_mounts = []
+    secret_env_vars = []
+    if (
+        resource_requirements is not None
+        and resource_requirements.kubernetes is not None
+    ):
         node_selector = resource_requirements.kubernetes.node_selector
         resource_requests = resource_requirements.kubernetes.requests
+        volume_info = _volume_secrets(resource_requirements.kubernetes.secret_mounts)
+        if volume_info is not None:
+            volume, mount = volume_info
+            volumes.append(volume)
+            volume_mounts.append(mount)
+        secret_env_vars.extend(
+            _environment_secrets(resource_requirements.kubernetes.secret_mounts)
+        )
         logger.debug("kubernetes node_selector %s", node_selector)
         logger.debug("kubernetes resource requests %s", resource_requests)
+        logger.debug("kubernetes volumes and mounts: %s, %s", volumes, volume_mounts)
+        logger.debug("kubernetes environment secrets: %s", secret_env_vars)
 
     job = kubernetes.client.V1Job(  # type: ignore
         api_version="batch/v1",
@@ -284,8 +305,9 @@ def _schedule_job(
                                     value=str(value),
                                 )
                                 for name, value in get_all_user_settings().items()
-                            ],
-                            volume_mounts=[],
+                            ]
+                            + secret_env_vars,
+                            volume_mounts=volume_mounts,
                             resources=(
                                 kubernetes.client.V1ResourceRequirements(  # type: ignore
                                     limits=resource_requests,
@@ -294,7 +316,7 @@ def _schedule_job(
                             ),
                         )
                     ],
-                    volumes=[],
+                    volumes=volumes,
                     tolerations=[],
                     restart_policy="Never",
                 ),
@@ -307,3 +329,87 @@ def _schedule_job(
     kubernetes.client.BatchV1Api().create_namespaced_job(  # type: ignore
         namespace=get_user_settings(SettingsVar.KUBERNETES_NAMESPACE), body=job
     )
+
+
+def _volume_secrets(
+    secret_mount: KubernetesSecretMount,
+) -> Optional[  # type: ignore
+    Tuple[kubernetes.client.V1Volume, kubernetes.client.V1VolumeMount]
+]:
+    """Configure a volume and corresponding mount for secrets requested for a func
+
+    Parameters
+    ----------
+    secret_mount:
+        The request for how to mount secrets into the pod for a Sematic func
+
+    Returns
+    -------
+    None if no file secrets were requested. Otherwise a volume and a volume mount
+    for the secrets requested.
+    """
+    if len(secret_mount.file_secrets) == 0:
+        return None
+
+    for relative_path in secret_mount.file_secrets.values():
+        if pathlib.Path(relative_path).is_absolute():
+            raise ValueError(
+                f"Cannot mount secret to absolute path '{relative_path}'; "
+                "paths must be relative."
+            )
+
+    volume_name = "sematic-func-secrets-volume"
+
+    volume = kubernetes.client.V1Volume(  # type: ignore
+        name=volume_name,
+        secret=kubernetes.client.V1SecretVolumeSource(  # type: ignore
+            items=[
+                kubernetes.client.V1KeyToPath(  # type: ignore
+                    key=key,
+                    path=relative_path,
+                )
+                for key, relative_path in secret_mount.file_secrets.items()
+            ],
+            optional=False,
+            secret_name=KUBERNETES_SECRET_NAME,
+        ),
+    )
+
+    mount = kubernetes.client.V1VolumeMount(  # type: ignore
+        mount_path=secret_mount.file_secret_root_path,
+        name=volume_name,
+        read_only=True,
+    )
+
+    return volume, mount
+
+
+def _environment_secrets(
+    secret_mount: KubernetesSecretMount,
+) -> List[kubernetes.client.V1EnvVar]:  # type: ignore
+    """Configure environment variables for secrets requested for a func
+
+    Parameters
+    ----------
+    secret_mount:
+        The request for how to mount secrets into the pod for a Sematic func
+
+    Returns
+    -------
+    A list of configurations for Kubernetes environment variables that will get
+    their values from the "sematic-func-secrets" Kubernetes secret.
+    """
+    env_vars = []
+    for key, env_var_name in secret_mount.environment_secrets.items():
+        env_vars.append(
+            kubernetes.client.V1EnvVar(  # type: ignore
+                name=env_var_name,
+                value_from=kubernetes.client.V1EnvVarSource(  # type: ignore
+                    secret_key_ref=kubernetes.client.V1SecretKeySelector(  # type: ignore
+                        name=KUBERNETES_SECRET_NAME,
+                        key=key,
+                    )
+                ),
+            )
+        )
+    return env_vars